fix SEGFAULT in GenomeGenerate

+++ src/utilsRefactored/seqFunctions.cpp
@@ -6,8 +6,7 @@ namespace RefactorProcessing {
 
     // encode a k-mer into an integer hash, return -1 if the seq length is less than kMerSize, or -i-2 if the i-th character is invalid
     int64_t encodeKmer(const std::string_view &seq, unsigned int kMerSize) {
-        // TODO: `kMerSize` should always be unsigned!
-        if (seq.length() < (size_t) kMerSize) {
+        if (seq.length() < kMerSize) {
             return -1;
         }
         int64_t hash = 0;
